+++ nft/model/topics.go
@@ -1,0 +1,7 @@
+package model
+
+import "github.com/ethereum/go-ethereum/common"
+
+var (
+	TransferTopicId = common.HexToHash("0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef")
+)